requirements.txt
altgraph==0.17.2
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==4.2.27
django-jazzmin==3.0.1
future==0.18.2
gunicorn==23.0.0
idna==3.11
macholib==1.15.2
packaging==25.0
Pillow==11.3.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.2
whitenoise==6.11.0
